Require readString to have an expected length
This patch adds a parameter to WrappedBuffer#readString which verifies that the string that we are reading is of the correct length. If the parsed length is larger than what is expected to be read, an error will be thrown. This has been introduced to prevent malicious servers consuming the user's memory.